# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sellaz Engine API",
        "description": ".",
        "contact": {
            "name": "Sellaz Support Team",
            "email": "support@sellaz.co.tz"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/orders/request": {
            "post": {
                "tags": [
                    "Super Dealer"
                ],
                "summary": "Place an order request",
                "description": "Place a new order as a Super Dealer.",
                "operationId": "aa9d67e66d91bbe87599169fa701e04a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "description": "ID of the product",
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the product",
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order placed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order placed successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "order_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only super dealers can place orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Only super dealers can place orders."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "products.0.product_id": [
                                                    "Each product must have a product ID."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Order creation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order creation failed."
                                        },
                                        "errors": {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Product ID 999 not found for this company."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get list of products",
                "description": "Fetches a list of all products of that particular company",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "List of products retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of all products"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new Product",
                "description": "Stores a new Product with a name, brand, company_id, and product image.",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "company_id",
                                    "brand",
                                    "image",
                                    "company_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Master Sports"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "SM"
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "company_price": {
                                        "type": "integer",
                                        "example": 12000
                                    },
                                    "image": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product added successfully"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to add Product."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product Model structure",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Master Sports"
                    },
                    "brand": {
                        "type": "string",
                        "example": "SM"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/images/logo.png"
                    },
                    "company_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "company_price": {
                        "type": "integer",
                        "example": 12000
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                },
                                "code": {
                                    "type": "integer",
                                    "example": 401
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Unprocessable Content",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Failed to validate fields"
                                },
                                "code": {
                                    "type": "integer",
                                    "example": 422
                                },
                                "errors": {
                                    "description": "Validation error messages, where each field contains an array of error messages.",
                                    "type": "object",
                                    "example": {
                                        "email": [
                                            "The email field is required."
                                        ],
                                        "password": [
                                            "The password must be at least 8 characters."
                                        ]
                                    },
                                    "additionalProperties": {
                                        "schema": "Controller",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "The field is required."
                                        }
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a valid bearer token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}